rename files and improve way of building IAlignedImages
